Adze 2.0.0
Light version less repetition.

+++ scripts/adze.sh
@@ -40,6 +40,8 @@ TNT_TAIL_FILE="cnRelayTestNet.log"
 #INITIALIZE ALL NECESSARY DIRECTORIES
 
         CETAK_PATH_TT=${CETAK_PATH}/${MAIN_FOLDER}
+        CETAK_PATH_SYS=${CETAK_PATH_TT}/cetak-sys
+        CETAK_PATH_SH=${CETAK_PATH_SYS}/SH
         CETAK_PATH_CNF=${CETAK_PATH_TT}/CONF
         CETAK_PATH_SNM=${CETAK_PATH_TT}/state-node-mainnet
         CETAK_PATH_SNT=${CETAK_PATH_TT}/state-node-testnet
@@ -62,6 +64,8 @@ TNT_TAIL_FILE="cnRelayTestNet.log"
 InitializeFolders() {   	#INITIALIZE ALL NECESSARY DIRECTORIES
 
         [[ -d ${CETAK_PATH_TT} ]] || mkdir -p -- ${CETAK_PATH}/${MAIN_FOLDER}
+        [[ -d ${CETAK_PATH_SYS} ]] || mkdir -p -- ${CETAK_PATH_TT}/cetak-sys
+        [[ -d ${CETAK_PATH_SH} ]] || mkdir -p -- ${CETAK_PATH_SYS}/SH
         [[ -d ${CETAK_PATH_CNF} ]] || mkdir -p -- ${CETAK_PATH_TT}/CONF
         [[ -d ${CETAK_PATH_SNM} ]] || mkdir -p -- ${CETAK_PATH_TT}/state-node-mainnet
         [[ -d ${CETAK_PATH_SNT} ]] || mkdir -p -- ${CETAK_PATH_TT}/state-node-testnet
@@ -77,7 +81,7 @@ InitializeFolders() {   	#INITIALIZE ALL NECESSARY DIRECTORIES
 ### -----------------------------
 
 # -------------------------------
-MAINNET() {                   	# MAINNET CONFIGURATION
+MAINNET(){                   	# MAINNET CONFIGURATION
 
 local vMnTOPOLOGY="$CETAK_PATH_CNF_MNT/mainnet-topology.json" # TOPOLOGY
 local vMnDATABASE="${CETAK_PATH_SNM}/db-mainnet" # DATABASE
@@ -85,6 +89,11 @@ local vMnSOCKET="${CETAK_PATH_SNM}/node.socket" # SOCKET
 
 local LAUNCHv2="${CETAK_PATH_TT}/${MNT_NODE_LINK}/bin/cardano-node"
 local vMnCONFIG="${CETAK_PATH_CNF_MNT}/mainnet-config.json" # CONF
+
+echo " LAUNCHv2: $LAUNCHv2"
+echo ""
+echo " vMnCONFIG: $vMnCONFIG"
+echo ""
 
 rm ${CETAK_PATH_TMP}/${MNT_TAIL_FILE}
 
@@ -111,6 +120,11 @@ local vTnSOCKET="${CETAK_PATH_SNT}/node.socket" # SOCKET
 local LAUNCHv2="${CETAK_PATH_TT}/${TNT_NODE_LINK}/bin/cardano-node"
 local vTnCONFIG="${CETAK_PATH_CNF_TNT}/testnet-config.json" # CONF
 
+echo " LAUNCHv2: $LAUNCHv2"
+echo ""
+echo " vTnCONFIG: $vTnCONFIG"
+echo ""
+
 rm ${CETAK_PATH_TMP}/${TNT_TAIL_FILE}
 
 nohup ${LAUNCHv2} run \
@@ -131,7 +145,7 @@ nohup ${LAUNCHv2} run \
 
 InstallCMD() {
 
-        cmdArray=( git wget tmux htop )
+        cmdArray=( curl jq git wget tmux htop )
         for vCmd in "${cmdArray[@]}"
         do
 	if ! command -v "$vCmd" > /dev/null 2>&1
@@ -147,76 +161,42 @@ InstallCMD() {
 }
 
 # -----------------------------------------------
-CNBuild() {					# INSTALL
+InstallMode() {                                 # INSTALL MODE
 
-                nix-collect-garbage -d
-                cd "${CETAK_PATH_TT}/cardano-node"
-                nix-build -A cardano-node -o ${CETAK_PATH_TT}/${MNT_NODE_LINK}
-                nix-build -A cardano-node -o ${CETAK_PATH_TT}/${TNT_NODE_LINK}
-                nix-build -A cardano-cli -o ${CETAK_PATH_TT}/cardano-cli-cetak
-                sleep 5
-                nix-collect-garbage -d
-
-}						# INSTALL
-# -----------------------------------------------
-
-# -----------------------------------------------
-InstallMN() {					# INSTALL MAINNET
-
-        # Default to Yes if the user presses enter without giving an answer:
-        echo " USER YOUR ATTENTION IS NEEDED "
+        echo " USER YOUR ATTENTION IS NEEDED " 
         echo
+	if ask "Should I Fetch configuration files?"; then # Only do something if you say Yes
+         case "$1" in
+                1)
+         	 rm -fv ${CETAK_PATH_CNF_MNT}/*.json
+                 cd $CETAK_PATH_CNF_MNT
+                 wget $mnnetConfig
+                 wget $mnnetByronGenesis
+                 wget $mnnetShelleyGenesis
+                 wget $mnnetTopology
+                ;;
 
-        # Only do something if you say Yes
-        if ask "Should I Fetch configuration files?"; then
-         rm -fv ${CETAK_PATH_CNF_MNT}/*.json
-        	cd $CETAK_PATH_CNF_MNT
-        	wget $mnnetConfig
-        	wget $mnnetByronGenesis
-        	wget $mnnetShelleyGenesis
-        	wget $mnnetTopology
+                2)
+         	rm -fv ${CETAK_PATH_CNF_TNT}/*.json
+                 cd $CETAK_PATH_CNF_TNT
+                 wget $ttnetConfig
+                 wget $ttnetByronGenesis
+                 wget $ttnetShelleyGenesis
+                 wget $ttnetTopology
+                ;;
+
+         *)
+            echo " Error $1 is'nt a good choice for InstallMode."
+            exit 1
+         esac
 	fi
-        ### Check if a directory does not exist ###
-        if [ ! -L "${CETAK_PATH_TT}/${MNT_NODE_LINK}" ] # IF LINK DON'T EXIST
-         then
-		CNBuild
-                exit # die with error code
-        fi
-}						# INSTALL MAINNET
-# -----------------------------------------------
-
-# -----------------------------------------------
-InstallTN() {					# INSTALL TESTNET
-
-	# Default to Yes if the user presses enter without giving an answer:
-	echo " USER YOUR ATTENTION IS NEEDED "
-	echo
-
-	# Only do something if you say Yes
-	if ask "Should I Fetch configuration files?"; then
-         rm -fv ${CETAK_PATH_CNF_TNT}/*.json
-                cd $CETAK_PATH_CNF_TNT
-                wget $ttnetConfig
-                wget $ttnetByronGenesis
-                wget $ttnetShelleyGenesis
-                wget $ttnetTopology
-	fi
-
-        ### Check if a directory does not exist ###
-
-        if [ ! -L "${CETAK_PATH_TT}/${TNT_NODE_LINK}" ] # IF LINK DON'T EXIST
-         then
-                CNBuild
-                exit # die with error code
-        fi
-
-
-}						# INSTALL TESTNET
+}                                               # INSTALL MODE
 # -----------------------------------------------
 
 # -----------------------------------------------
 getCN() {					# NODE
 
+	rm -rf $CETAK_PATH_TT/cardano-node
 	cd $CETAK_PATH_TT
 	git clone --recurse-submodules $CN_GIT_WEB
 	cd $CETAK_PATH_TT/cardano-node
@@ -230,45 +210,49 @@ getCN() {					# NODE
 # -----------------------------------------------
 
 # -----------------------------------------------
-NodeEnvnmt() {					# NODE ENVIRONEMENT
-
-        InitializeFolders
-        InstallCMD
-
-}						# NODE ENVIRONEMENT
-# -----------------------------------------------
-
-# -----------------------------------------------
 NodeMaintenance() {                             # NODE
 
         # Only do something if you say Yes
-        if ask "Attention this will update the Node any changes will permanently be removed!"; then
+        if ask "Attention this process will install the Node and remove any previous installation!"; then
+	 if [ $1 -lt 2 ]
+	 then
+          InitializeFolders
+          InstallCMD
+          pkill cardano-node -SIGINT
+          getCN
+	 fi
 	 case "$1" in
         	1)
-		 pkill cardano-node -SIGINT
-                 rm -rf $CETAK_PATH_TT/cardano-node
                  rm -rf ${CETAK_PATH_TT}/${MNT_NODE_LINK}
-                 getCN
-            	 InstallMN
+		 InstallMode 1
+		 cd ${CETAK_PATH_TT}/cardano-node
+                 nix-build -A cardano-node -o ${CETAK_PATH_TT}/${MNT_NODE_LINK} # INSTALL MAINNET
+                 #exit
             	;;
 
         	2)
-                 pkill cardano-node -SIGINT
-		 rm -rf $CETAK_PATH_TT/cardano-node
          	 rm -rf ${CETAK_PATH_TT}/${TNT_NODE_LINK}
-		 getCN
-            	 InstallTN
+                 InstallMode 2
+                 cd ${CETAK_PATH_TT}/cardano-node
+                 nix-build -A cardano-node -o ${CETAK_PATH_TT}/${TNT_NODE_LINK} # INSTALL TESTNET
+                 #exit
             	;;
 
-        	9)
-            	echo " This is Optional Node Maintenance "
-            	;;
+                9)
+		 echo " Test Option "
+                 exit
+                ;;
 
-         *)
+            *)
             echo " Error $1 is'nt a good choice."
             exit 1
-
 	 esac
+         if ask "Should I install Cardano-cli?"; then # Only do something if you say Yes
+          nix-build -A cardano-cli -o ${CETAK_PATH_TT}/cardano-cli-cetak
+         fi
+	 if ask "A healthy machine needs a cleen up should I?"; then # Only do something if you say Yes
+          nix-collect-garbage -d
+	 fi
 	fi
 }						# NODE
 # -----------------------------------------------
@@ -350,50 +334,45 @@ do
 	echo "-------------------------------"
 	echo "     M A I N - M E N U"
 	echo "-------------------------------"
-	echo "1. Install Default Command."
-	echo "2. Install or Update MainNet."
-	echo "3. Install or Update TestNet."
-        echo "4. Optional install mode."
-        echo "5. Default Tail."
-        echo "6. Launch MAINnet Node."
-        echo "7. Launch TESTnet Node."
-        echo "8. SHUTDOWN"
+	echo "1. Install MainNet."
+	echo "2. Install TestNet."
+        echo "3. Optional install mode."
+        echo "4. Default Tail."
+        echo "5. Launch MAINnet Node."
+        echo "6. Launch TESTnet Node."
+        echo "7. SHUTDOWN"
 	echo "9. Exit"
         # get input from the user
 	read -p "Enter your choice [ 1 - 7 ] " choice
         # make decision using case..in..esac
 	case $choice in
 		1)
-			ask "Do you want to install suggested commands?" && NodeEnvnmt # Only do something if you say Yes
+			ask "Install Mainnet?" && NodeMaintenance 1 # Only do something if you say Yes
 			read -p "Press [Enter] key to continue..." readEnterKey
 			;;
 		2)
-			ask "Install/Update Mainnet?" && NodeMaintenance 1 # Only do something if you say Yes
+			ask "Install TestNet?" && NodeMaintenance 2 # Only do something if you say Yes
 			read -p "Press [Enter] key to continue..." readEnterKey
 			;;
-		3)
-			ask "Install/Update TestNet?" && NodeMaintenance 2 # Only do something if you say Yes
-			read -p "Press [Enter] key to continue..." readEnterKey
-			;;
-                4)
+                3)
                         ask "This is the optional install mode!" && NodeMaintenance 9 # Only do something if you say Yes
                         read -p "Press [Enter] key to continue..." readEnterKey
                         ;;
-                5)
+                4)
                         ask "Default Tail?" && MultiChoiceNode 1 # Only do something if you say Yes
                         read -p "Press [Enter] key to continue..." readEnterKey
                         ;;
-                6)
+                5)
                         ask "Launch MAINnet Node?" && MAINNET # Only do something if you say Yes
 			exit
                         #read -p "Press [Enter] key to continue..." readEnterKey
                         ;;
-                7)
+                6)
                         ask "Launch TESTnet Node?" && TESTNET # Only do something if you say Yes
                         exit
                         #read -p "Press [Enter] key to continue..." readEnterKey
                         ;;
-                8)
+                7)
                         ask "SHUTDOWN" && MultiChoiceNode 2 # Only do something if you say Yes
                         exit
                         #read -p "Press [Enter] key to continue..." readEnterKey
